Enable the clang-diagnostic-format checks
Re-enables clang-diagnostic-format, -format-nonliteral, -format-security
and -format-signedness in .clang-tidy and clears the ~160 resulting
errors across ra and td.

Format strings that are only known at runtime are the bulk of them: the
localized string table, or a format forwarded from a caller. The
compiler cannot check those, so rather than repeat the suppression at
every print site, jshell.h gains

  void Format_Runtime_Text(char* buffer, size_t size, const char* fmt, ...);
  void Format_Runtime_Text(char* buffer, size_t size, const char* fmt, va_list);

which wraps vsnprintf with the suppression in one place, in both ports.
Roughly fifty sprintf/vsprintf calls now go through it, so they also
truncate instead of overflowing their fixed buffers. Where the format
was a plain string with no arguments, snprintf("%s", ...) does instead.
Note that MonoClass::Printf loses the comment about placing its buffer
last in the frame "so that if the sprintf happens to spill past the end,
it isn't likely to trash anything" -- with vsnprintf it cannot spill.

Several of the diagnostics were real defects rather than cosmetic:

- td/jshell.cc passed Fatal's caller-supplied message to Mono_Printf as
  the format, so a crash message containing a % read garbage off the
  stack. td/conquer.cc (CD prompts) and both netdlg.cc files (the
  "waiting for connections" banner) had the same shape with translated
  strings.
- td/mplayer.cc read the SyncBug "this" trap with sscanf("%x", &TrapThis)
  into a void*, writing four bytes and corrupting the adjacent four on
  any 64-bit build. It now parses to uintptr_t via SCNxPTR.
- ra/queue.cc printed xTargetClass objects through %x, passing a
  non-trivial class through varargs. The surrounding "Whom:%x (%x)"
  format wants the handle and its id, so it prints As_TARGET() and
  Value().
- Both mapedit.cc files had printf("%s\n", txt) where txt is an int
  string-table index, which would have crashed the map editor's fatal
  handler.

Two type mismatches are fixed at the root instead of at the call site,
because the variable disagreed with what fed it: LogLevelTime and
LogLastTime are int64_t now, matching TickCount.Value(), and queue.cc's
their_frame[] is int64_t, matching Frame. Everywhere else the type is
pinned by a serialized struct (TEventClass::Data.Value, Session.MaxAhead)
or by a caller, so the cast goes at the printf.

Also in passing: ra/unit.cc reached UnitTypeClass::Harvester_Load_List
and _Dump_List through Class->, which readability-static-accessed-through
-instance flags; two lines nearby already spelled it out. td links
absl::core_headers explicitly now that its jshell.h uses
ABSL_PRINTF_ATTRIBUTE.

Verified by compiling every ra and td translation unit with the format
warnings enabled -- note the local build dir sets STRICT_CHECKS=OFF,
which puts -w in every compile command and silently suppresses them --
and by diffing each changed file's full -Wall -Wextra -Wconversion
-Wsign-conversion warning set against its previous version.

+++ src/ra/radar.cc
@@ -2261,7 +2261,7 @@ bool RadarClass::Draw_House_Info() {
     y += 12 + 1;
 
     // count & print buildings
-    snprintf(txt, sizeof(txt), "%i", ptr->CurBuildings);
+    snprintf(txt, sizeof(txt), "%i", static_cast<int>(ptr->CurBuildings));
     Fancy_Text_Print(txt, RadX + RadOffX + 12, y, color, BLACK,
                      style);
     y += 12 + 1;
@@ -2271,7 +2271,7 @@ bool RadarClass::Draw_House_Info() {
                      TPF_6PT_GRAD | TPF_NOSHADOW);
     y += 12 + 1;
     // count & print units
-    snprintf(txt, sizeof(txt), "%i", ptr->CurUnits);
+    snprintf(txt, sizeof(txt), "%i", static_cast<int>(ptr->CurUnits));
     Fancy_Text_Print(txt, RadX + RadOffX + 12, y, color, BLACK,
                      style);
     y += 12 + 1;
@@ -2281,7 +2281,7 @@ bool RadarClass::Draw_House_Info() {
                      TPF_6PT_GRAD | TPF_NOSHADOW);
     y += 12 + 1;
     // count & print infantry
-    snprintf(txt, sizeof(txt), "%i", ptr->CurInfantry);
+    snprintf(txt, sizeof(txt), "%i", static_cast<int>(ptr->CurInfantry));
     Fancy_Text_Print(txt, RadX + RadOffX + 12, y, color, BLACK,
                      style);
 #if (0)
