+++ src/td/building.cc
@@ -3132,7 +3132,8 @@ void BuildingClass::Write_INI(char* buffer) {
     if (!building->IsInLimbo) {
       sprintf(uname, "%03d", index);
       sprintf(buf, "%s,%s,%d,%u,%d,%s", building->House->Class->IniName,
-              building->Class->IniName, building->Health_Ratio(),
+              building->Class->IniName,
+              static_cast<int>(building->Health_Ratio()),
               Coord_Cell(building->Coord), building->PrimaryFacing.Current(),
               building->Trigger ? building->Trigger->Get_Name() : "None");
       WWWritePrivateProfileString(INI_Name(), uname, buf, buffer);
